Rename RHTAP to Konflux CI
Note that the thing that was previously called RHTAP, is almost, but
not entirely the same as the thing that is now called Konflux, so
there may be further revisions needed. This commit is mostly just an
unsophisticated search and replace.

(We might also want to consider splitting Konflux specific docs into
another location but that can come later.)

Ref: https://issues.redhat.com/browse/EC-378